+++ test/integration/expression/expression.test.ts
@@ -12,7 +12,6 @@ import {
 import ExpressionParsingError from '../../../src/expression/parsing_error';
 import {Result} from '../../../src/util/result';
 import {getGeometry} from '../../lib/geometry';
-import {stringify} from '../../lib/util';
 import {deepEqual, stripPrecision} from '../../lib/json-diff';
 
 const DECIMAL_SIGNIFICANT_FIGURES =  6;
@@ -52,7 +51,7 @@ describe('expression', () => {
                 delete fixture.metadata;
 
                 const fname = path.join(__dirname, expressionTestFileName);
-                fs.writeFileSync(fname, `${stringify(fixture)}\n`);
+                fs.writeFileSync(fname, JSON.stringify(fixture));
                 return;
             }